Add `predict_data` to `thunder run`. Move `last_checkpoint` to torch.utils
